+++ dst/app/middlewares/authentication.js
@@ -9,32 +9,41 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.authentication = void 0;
 /**
  * OAuthミドルウェア
- * @see https://aws.amazon.com/blogs/mobile/integrating-amazon-cognito-user-pools-with-api-gateway/
  */
 const express_middleware_1 = require("@motionpicture/express-middleware");
 const http_status_1 = require("http-status");
 const api_1 = require("../error/api");
 // 許可発行者リスト
 const ISSUERS = process.env.TOKEN_ISSUERS.split(',');
+const TOKEN_ISSUER_REQUEST_TIMEOUT = 5000;
 // tslint:disable-next-line:no-single-line-block-comment
 /* istanbul ignore next */
-exports.default = (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
-    try {
-        yield (0, express_middleware_1.cognitoAuth)({
-            issuers: ISSUERS,
-            authorizedHandler: (user, token) => __awaiter(void 0, void 0, void 0, function* () {
-                req.user = user;
-                req.accessToken = token;
-                next();
-            }),
-            unauthorizedHandler: (err) => {
-                next(new api_1.APIError(http_status_1.UNAUTHORIZED, [new Error(err.message)]));
-            }
-        })(req, res, next);
-    }
-    catch (error) {
-        next(new api_1.APIError(http_status_1.UNAUTHORIZED, [new Error(error.message)]));
-    }
-});
+function authentication(req, res, next) {
+    return __awaiter(this, void 0, void 0, function* () {
+        try {
+            yield (0, express_middleware_1.cognitoAuth)({
+                authorizedHandler: (user, token) => __awaiter(this, void 0, void 0, function* () {
+                    req.user = user;
+                    req.accessToken = token;
+                    next();
+                }),
+                unauthorizedHandler: (err) => {
+                    next(new api_1.APIError(http_status_1.UNAUTHORIZED, [new Error(err.message)]));
+                },
+                requestOptions: { timeout: TOKEN_ISSUER_REQUEST_TIMEOUT },
+                verifyOptions: {
+                    tokenUse: 'access',
+                    decodeWithoutVerifying: false,
+                    issuers: ISSUERS
+                }
+            })(req, res, next);
+        }
+        catch (error) {
+            next(new api_1.APIError(http_status_1.UNAUTHORIZED, [new Error(error.message)]));
+        }
+    });
+}
+exports.authentication = authentication;