+++ src/app/middlewares/authentication.ts
@@ -1,6 +1,5 @@
 /**
  * OAuthミドルウェア
- * @see https://aws.amazon.com/blogs/mobile/integrating-amazon-cognito-user-pools-with-api-gateway/
  */
 import { cognitoAuth } from '@motionpicture/express-middleware';
 import { NextFunction, Request, Response } from 'express';
@@ -10,13 +9,13 @@ import { APIError } from '../error/api';
 
 // 許可発行者リスト
 const ISSUERS = (<string>process.env.TOKEN_ISSUERS).split(',');
+const TOKEN_ISSUER_REQUEST_TIMEOUT: number = 5000;
 
 // tslint:disable-next-line:no-single-line-block-comment
 /* istanbul ignore next */
-export default async (req: Request, res: Response, next: NextFunction) => {
+export async function authentication(req: Request, res: Response, next: NextFunction) {
     try {
         await cognitoAuth({
-            issuers: ISSUERS,
             authorizedHandler: async (user, token) => {
                 req.user = user;
                 req.accessToken = token;
@@ -24,9 +23,15 @@ export default async (req: Request, res: Response, next: NextFunction) => {
             },
             unauthorizedHandler: (err) => {
                 next(new APIError(UNAUTHORIZED, [new Error(err.message)]));
+            },
+            requestOptions: { timeout: TOKEN_ISSUER_REQUEST_TIMEOUT },
+            verifyOptions: {
+                tokenUse: 'access',
+                decodeWithoutVerifying: false,
+                issuers: ISSUERS
             }
         })(req, res, next);
     } catch (error) {
         next(new APIError(UNAUTHORIZED, [new Error(error.message)]));
     }
-};
+}